Remove Setup tests and update OSX Peer trust failures to the correct issue * The setup tests do not add much value today as we output certification installation trace that help identify certificate setup failures. * The set up tests are also fragile because it does not use the same way we install the certificates on Linux/Osx machines, and it's non trial to make them the same way.
